value: eliminate some casts from Word
We were storing Word::n as a u32 even though it fits into a u8. u8 can
be converted to every unsigned integer type without casts, while u32
cannot. So by using u8, we can eliminate casts everywhere.

+++ src/bit_encoding/encode.rs
@@ -258,7 +258,7 @@ fn encode_node<N: node::Marker>(
             node::Inner::Word(word) => {
                 w.write_bit(true)?; // jet or word
                 w.write_bit(false)?; // word
-                encode_natural(1 + word.n() as usize, w)?;
+                encode_natural(1 + word.n(), w)?;
                 encode_value(word.as_value(), w)?;
             }
             _ => unreachable!(),
